Adjust code to work for CI and Local.
